+++ tests/utils.py
@@ -351,6 +351,9 @@ class Utils:
                 raise TypeError(
                     "input_data must be a DataFrame, a list of Row objects or a Row object"
                 )
+
+            # Strip column names to make errors more concise
+            rows = [Row(*list(x)) for x in rows]
             return rows
 
         actual_rows = get_rows(actual)
@@ -382,6 +385,7 @@ class Utils:
                     meta.precision,
                     meta.scale,
                     meta.internal_size,
+                    session._conn.max_string_size,
                 )
                 == field.datatype
             )
@@ -1159,6 +1163,21 @@ class TestData:
                 cls.Number2(2, 1, 10),
                 cls.Number2(2, 2, 1),
                 cls.Number2(2, 2, 3),
+            ]
+        )
+
+    @classmethod
+    def xyz2(cls, session: "Session") -> DataFrame:
+        return session.create_dataframe(
+            [
+                cls.Number2(1, 2, 1),
+                cls.Number2(1, 2, 3),
+                cls.Number2(2, 1, 10),
+                cls.Number2(2, 2, 1),
+                cls.Number2(2, 2, 3),
+                cls.Number2(2, 3, 5),
+                cls.Number2(2, 3, 8),
+                cls.Number2(2, 4, 7),
             ]
         )